refactor: relocate frontend tests and add PostCSS and Tailwind resolve aliases to vite.config.ts
